refactor: Remove strict validation test from `NestedSetsQueryBehaviorTest` class and clean up imports.
