+++ tests/NestedSetsBehaviorTest.php
@@ -10,7 +10,7 @@ use yii\base\NotSupportedException;
 use yii\db\{ActiveRecord, Exception, StaleObjectException};
 use yii\helpers\ArrayHelper;
 use yii2\extensions\nestedsets\NestedSetsBehavior;
-use yii2\extensions\nestedsets\tests\support\model\{MultipleTree, Tree};
+use yii2\extensions\nestedsets\tests\support\model\{MultipleTree, Tree, TreeWithStrictValidation};
 
 use function get_class;
 use function sprintf;
@@ -1781,5 +1781,53 @@ final class NestedSetsBehaviorTest extends TestCase
         $this->expectExceptionMessage('Value cannot be \'null\' in \'beforeInsertNode()\' method.');
 
         $childNode->appendTo($parentNode);
+    }
+
+    public function testAppendToWithRunValidationParameterUsingStrictValidation(): void
+    {
+        $this->generateFixtureTree();
+
+        $targetNode = Tree::findOne(2);
+
+        self::assertNotNull(
+            $targetNode,
+            'Target node with ID \'2\' should exist before calling \'appendTo\'.',
+        );
+
+        $invalidNode = new TreeWithStrictValidation(['name' => 'x']);
+
+        $result1 = $invalidNode->appendTo($targetNode);
+        $hasError1 = $invalidNode->hasErrors();
+
+        self::assertFalse(
+            $result1,
+            '\'appendTo()\' should return \'false\' when \'runValidation=true\' and data fails validation.',
+        );
+        self::assertTrue(
+            $hasError1,
+            'Node should have validation errors when \'runValidation=true\' and data is invalid.',
+        );
+
+        $invalidNode2 = new TreeWithStrictValidation(['name' => 'x']);
+
+        $result2 = $invalidNode2->appendTo($targetNode, false);
+        $hasError2 = $invalidNode2->hasErrors();
+
+        self::assertTrue(
+            $result2,
+            '\'appendTo()\' should return \'true\' when \'runValidation=false\', even with invalid data ' .
+            'that would fail validation.',
+        );
+        self::assertFalse(
+            $hasError2,
+            'Node should not have validation errors when \'runValidation=false\' because validation was skipped.',
+        );
+
+        $persistedNode = TreeWithStrictValidation::findOne($invalidNode2->id);
+
+        self::assertNotNull(
+            $persistedNode,
+            'Node should exist in database after appending to target node with validation disabled.',
+        );
     }
 }
